% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractPropertyValue.r
\docType{data}
\name{AbstractPropertyValue}
\alias{AbstractPropertyValue}
\title{AbstractPropertyValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractPropertyValue
}
\description{
AbstractPropertyValue Class
}
\section{Fields}{

\describe{
\item{\code{pred}}{}

\item{\code{xrefs}}{}

\item{\code{val}}{}
}}

\keyword{datasets}
