% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationExtension.r
\docType{data}
\name{AnnotationExtension}
\alias{AnnotationExtension}
\title{AnnotationExtension Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AnnotationExtension
}
\description{
AnnotationExtension Class
}
\section{Fields}{

\describe{
\item{\code{filler}}{}

\item{\code{relation_chain}}{}
}}

\keyword{datasets}
