% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APageOfResults.r
\docType{data}
\name{APageOfResults}
\alias{APageOfResults}
\title{APageOfResults Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
APageOfResults
}
\description{
APageOfResults Class
}
\section{Fields}{

\describe{
\item{\code{total}}{}

\item{\code{page}}{}

\item{\code{pages}}{}

\item{\code{per_page}}{}
}}

\keyword{datasets}
