% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Association.r
\docType{data}
\name{Association}
\alias{Association}
\title{Association Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Association
}
\description{
Association Class
}
\section{Fields}{

\describe{
\item{\code{subject}}{}

\item{\code{object}}{}

\item{\code{id}}{}

\item{\code{slim}}{}

\item{\code{provided_by}}{}

\item{\code{subject_extensions}}{}

\item{\code{publications}}{}

\item{\code{qualifiers}}{}

\item{\code{object_extensions}}{}

\item{\code{evidence_graph}}{}

\item{\code{negated}}{}

\item{\code{evidence_types}}{}

\item{\code{type}}{}

\item{\code{relation}}{}
}}

\keyword{datasets}
