% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationApi.r
\docType{data}
\name{AssociationApi}
\alias{AssociationApi}
\title{Association operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AssociationApi
}
\description{
swagger.Association
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_association_by_subject_and_object_category_search Returns list of matching associations


get_association_by_subject_category_search Returns list of matching associations


get_association_object Returns the association with a given identifier


get_association_search Generalized search over complete corpus of associations


get_associations_between Returns associations connecting two entities


get_associations_from Returns list of matching associations starting from a given subject (source)


get_associations_to Returns list of matching associations pointing to a given object (target)

}
}

\keyword{datasets}
