% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssociationResults.r
\docType{data}
\name{AssociationResults}
\alias{AssociationResults}
\title{AssociationResults Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AssociationResults
}
\description{
AssociationResults Class
}
\section{Fields}{

\describe{
\item{\code{start}}{}

\item{\code{facet_pivot}}{}

\item{\code{facet_counts}}{}

\item{\code{numFound}}{}

\item{\code{objects}}{}

\item{\code{associations}}{}

\item{\code{compact_associations}}{}
}}

\keyword{datasets}
