% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioentityApi.r
\docType{data}
\name{BioentityApi}
\alias{BioentityApi}
\title{Bioentity operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioentityApi
}
\description{
swagger.Bioentity
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_allele_object Returns genotype object


get_anatomy_gene_associations Returns expression events for a gene


get_anatomy_gene_by_taxon_associations Returns gene ids for all genes for a particular anatomy in a taxon


get_anatomy_object Returns basic info on anatomical entity


get_disease_gene_associations Returns genes associated with a disease


get_disease_model_associations Returns associations to models of the disease


get_disease_model_taxon_associations Returns associations to models of the disease constrained by taxon


get_disease_object Returns basic info on a disease


get_disease_phenotype_associations Returns phenotypes associated with disease


get_disease_substance_associations Returns substances associated with a disease


get_gene_disease_associations Returns diseases associated with gene


get_gene_expression_associations Returns expression events for a gene


get_gene_function_associations Returns function associations for a gene


get_gene_homolog_associations Returns homologs for a gene


get_gene_interactions Returns interactions for a gene


get_gene_object Returns basic info about a gene


get_gene_pathway_associations Returns pathways associated with gene


get_gene_phenotype_associations Returns phenotypes associated with gene


get_generic_associations Returns associations for an entity regardless of the type


get_generic_object Returns basic info on object of any type


get_genotype_disease_associations Returns diseases associated with a genotype


get_genotype_gene_associations Returns genes associated with a genotype


get_genotype_genotype_associations Returns genotypes-genotype associations


get_genotype_object Returns genotype object


get_genotype_phenotype_associations Returns phenotypes associated with a genotype


get_goterm_gene_associations Returns associations to GO terms for a gene


get_goterm_object Returns basic info on a gene


get_parent_object Returns basic info on an individual/case


get_pathway_gene_associations Returns genes associated with a pathway


get_pathway_object Returns basic info on a pathway


get_phenotype_anatomy_associations Returns anatomical entities associated with a phenotype


get_phenotype_function_associations Returns biological functions associated with a Phenotype


get_phenotype_gene_associations Returns associated phenotypes


get_phenotype_gene_by_taxon_associations Returns gene ids for all genes for a particular phenotype in a taxon


get_phenotype_object Returns basic info on a phenotype


get_substance_object Returns basic info on a substance


get_substance_participant_in_associations Returns associations between an activity and process and the specified substance


get_substance_role_associations Returns associations between given drug and roles


get_substance_treats_associations Returns substances associated with a disease


get_variant_gene_associations Returns genes associated with a variant


get_variant_genotype_associations Returns genotypes associated with a variant


get_variant_object Returns basic info on a variant


get_variant_phenotype_associations Returns phenotypes associated with a variant

}
}

\keyword{datasets}
