% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioentitysetApi.r
\docType{data}
\name{BioentitysetApi}
\alias{BioentitysetApi}
\title{Bioentityset operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioentitysetApi
}
\description{
swagger.Bioentityset
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_entity_set_associations Returns compact associations for a given input set


get_entity_set_graph_resource TODO Graph object spanning all entities


get_entity_set_summary Summary statistics for objects associated


get_over_representation Summary statistics for objects associated

}
}

\keyword{datasets}
