% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioentitysethomologsApi.r
\docType{data}
\name{BioentitysethomologsApi}
\alias{BioentitysethomologsApi}
\title{Bioentitysethomologs operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioentitysethomologsApi
}
\description{
swagger.Bioentitysethomologs
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_entity_set_homologs Returns homology associations for a given input set of genes

}
}

\keyword{datasets}
