% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioentitysetslimmerApi.r
\docType{data}
\name{BioentitysetslimmerApi}
\alias{BioentitysetslimmerApi}
\title{Bioentitysetslimmer operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioentitysetslimmerApi
}
\description{
swagger.Bioentitysetslimmer
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_entity_set_slimmer Summarize a set of objects

}
}

\keyword{datasets}
