% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioObject.r
\docType{data}
\name{BioObject}
\alias{BioObject}
\title{BioObject Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioObject
}
\description{
BioObject Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}
}}

\keyword{datasets}
