% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioObjectCore.r
\docType{data}
\name{BioObjectCore}
\alias{BioObjectCore}
\title{BioObjectCore Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BioObjectCore
}
\description{
BioObjectCore Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{taxon}}{}
}}

\keyword{datasets}
