% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CamApi.r
\docType{data}
\name{CamApi}
\alias{CamApi}
\title{Cam operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CamApi
}
\description{
swagger.Cam
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_activity_collection Returns list of models


get_instance_object Returns list of matches


get_model_collection Returns list of ALL models


get_model_contibutors Returns list of all contributors across all models


get_model_instances Returns list of all instances


get_model_object Returns a complete model


get_model_properties Returns list of all properties used across all models


get_model_property_values Returns list property-values for all models


get_model_query Returns list of models matching query


get_physical_interaction Returns list of models

}
}

\keyword{datasets}
