% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromosome.r
\docType{data}
\name{Chromosome}
\alias{Chromosome}
\title{Chromosome Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Chromosome
}
\description{
Chromosome Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}
}}

\keyword{datasets}
