% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompactAssociationSet.r
\docType{data}
\name{CompactAssociationSet}
\alias{CompactAssociationSet}
\title{CompactAssociationSet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CompactAssociationSet
}
\description{
CompactAssociationSet Class
}
\section{Fields}{

\describe{
\item{\code{objects}}{}

\item{\code{subject}}{}

\item{\code{relation}}{}
}}

\keyword{datasets}
