% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Edge.r
\docType{data}
\name{Edge}
\alias{Edge}
\title{Edge Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Edge
}
\description{
Edge Class
}
\section{Fields}{

\describe{
\item{\code{sub}}{}

\item{\code{pred}}{}

\item{\code{obj}}{}
}}

\keyword{datasets}
