% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntityReference.r
\docType{data}
\name{EntityReference}
\alias{EntityReference}
\title{EntityReference Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EntityReference
}
\description{
EntityReference Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{categories}}{}

\item{\code{label}}{}
}}

\keyword{datasets}
