% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvidencegraphApi.r
\docType{data}
\name{EvidencegraphApi}
\alias{EvidencegraphApi}
\title{Evidencegraph operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EvidencegraphApi
}
\description{
swagger.Evidencegraph
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_evidence_graph_image Returns evidence graph as a png


get_evidence_graph_object Returns evidence graph object for a given association

}
}

\keyword{datasets}
