% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene.r
\docType{data}
\name{Gene}
\alias{Gene}
\title{Gene Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Gene
}
\description{
Gene Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{literature_associations}}{}

\item{\code{summaries}}{}

\item{\code{disease_associations}}{}

\item{\code{full_name}}{}

\item{\code{phenotype_associations}}{}

\item{\code{families}}{}

\item{\code{function_associations}}{}

\item{\code{chromosome}}{}

\item{\code{genotype_associations}}{}

\item{\code{systematic_name}}{}

\item{\code{sequence_features}}{}

\item{\code{homology_associations}}{}

\item{\code{interaction_associations}}{}

\item{\code{transcripts}}{}

\item{\code{pathway_associations}}{}
}}

\keyword{datasets}
