% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomefeaturesApi.r
\docType{data}
\name{GenomefeaturesApi}
\alias{GenomefeaturesApi}
\title{Genomefeatures operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GenomefeaturesApi
}
\description{
swagger.Genomefeatures
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_features_within_resource Returns list of matches

}
}

\keyword{datasets}
