% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype.r
\docType{data}
\name{Genotype}
\alias{Genotype}
\title{Genotype Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Genotype
}
\description{
Genotype Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{phenotype_associations}}{}

\item{\code{gene_associations}}{}

\item{\code{disease_associations}}{}

\item{\code{variant_associations}}{}
}}

\keyword{datasets}
