% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.r
\docType{data}
\name{Graph}
\alias{Graph}
\title{Graph Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Graph
}
\description{
Graph Class
}
\section{Fields}{

\describe{
\item{\code{edges}}{}

\item{\code{nodes}}{}
}}

\keyword{datasets}
