% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphApi.r
\docType{data}
\name{GraphApi}
\alias{GraphApi}
\title{Graph operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GraphApi
}
\description{
swagger.Graph
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_edge_resource Returns edges emanating from a node


get_node_resource Returns a graph node

}
}

\keyword{datasets}
