% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifiermapperApi.r
\docType{data}
\name{IdentifiermapperApi}
\alias{IdentifiermapperApi}
\title{Identifiermapper operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IdentifiermapperApi
}
\description{
swagger.Identifiermapper
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_identifier_mapper TODO maps a list of identifiers from a source to a target

}
}

\keyword{datasets}
