% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentifierprefixesApi.r
\docType{data}
\name{IdentifierprefixesApi}
\alias{IdentifierprefixesApi}
\title{Identifierprefixes operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IdentifierprefixesApi
}
\description{
swagger.Identifierprefixes
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_prefix_collection Returns list of prefixes


get_prefix_contract Returns contracted URI


get_prefix_expand Returns expanded URI

}
}

\keyword{datasets}
