% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndividualApi.r
\docType{data}
\name{IndividualApi}
\alias{IndividualApi}
\title{Individual operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
IndividualApi
}
\description{
swagger.Individual
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_individual Returns list of matches


get_pedigree Returns list of matches

}
}

\keyword{datasets}
