% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MartApi.r
\docType{data}
\name{MartApi}
\alias{MartApi}
\title{Mart operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MartApi
}
\description{
swagger.Mart
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_mart_case_associations_resource Bulk download of case associations


get_mart_disease_associations_resource Bulk download of disease associations


get_mart_gene_associations_resource Bulk download of gene associations

}
}

\keyword{datasets}
