% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NamedObjectCore.r
\docType{data}
\name{NamedObjectCore}
\alias{NamedObjectCore}
\title{NamedObjectCore Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NamedObjectCore
}
\description{
NamedObjectCore Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}
}}

\keyword{datasets}
