% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NlpannotateApi.r
\docType{data}
\name{NlpannotateApi}
\alias{NlpannotateApi}
\title{Nlpannotate operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NlpannotateApi
}
\description{
swagger.Nlpannotate
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_annotate Not yet implemented

}
}

\keyword{datasets}
