% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Node.r
\docType{data}
\name{Node}
\alias{Node}
\title{Node Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Node
}
\description{
Node Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{lbl}}{}
}}

\keyword{datasets}
