% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntolApi.r
\docType{data}
\name{OntolApi}
\alias{OntolApi}
\title{Ontol operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OntolApi
}
\description{
swagger.Ontol
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_extract_ontology_subgraph_resource Extract a subgraph from an ontology


get_information_content_resource Returns information content (IC) for a set of relevant ontology classes


get_test_me TEST

}
}

\keyword{datasets}
