% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OntollabelerApi.r
\docType{data}
\name{OntollabelerApi}
\alias{OntollabelerApi}
\title{Ontollabeler operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OntollabelerApi
}
\description{
swagger.Ontollabeler
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_ontol_labeler_resource Fetches a map from CURIEs/IDs to labels

}
}

\keyword{datasets}
