% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OwlontologyApi.r
\docType{data}
\name{OwlontologyApi}
\alias{OwlontologyApi}
\title{Owlontology operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
OwlontologyApi
}
\description{
swagger.Owlontology
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_dl_query Placeholder - use OWLery for now


get_sparql_query Placeholder - use direct SPARQL endpoint for now

}
}

\keyword{datasets}
