% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageOfVariantSets.r
\docType{data}
\name{PageOfVariantSets}
\alias{PageOfVariantSets}
\title{PageOfVariantSets Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PageOfVariantSets
}
\description{
PageOfVariantSets Class
}
\section{Fields}{

\describe{
\item{\code{total}}{}

\item{\code{page}}{}

\item{\code{pages}}{}

\item{\code{per_page}}{}

\item{\code{items}}{}
}}

\keyword{datasets}
