% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Publication.r
\docType{data}
\name{Publication}
\alias{Publication}
\title{Publication Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Publication
}
\description{
Publication Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}
}}

\keyword{datasets}
