% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubpubsApi.r
\docType{data}
\name{PubpubsApi}
\alias{PubpubsApi}
\title{Pubpubs operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PubpubsApi
}
\description{
swagger.Pubpubs
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_foo TODO Returns list of matches

}
}

\keyword{datasets}
