% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relation.r
\docType{data}
\name{Relation}
\alias{Relation}
\title{Relation Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Relation
}
\description{
Relation Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}
}}

\keyword{datasets}
