% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationRef.r
\docType{data}
\name{RelationRef}
\alias{RelationRef}
\title{RelationRef Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RelationRef
}
\description{
RelationRef Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}
}}

\keyword{datasets}
