% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelationusageApi.r
\docType{data}
\name{RelationusageApi}
\alias{RelationusageApi}
\title{Relationusage operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RelationusageApi
}
\description{
swagger.Relationusage
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_relation_usage_between_resource All relations used plus count of associations


get_relation_usage_pivot_label_resource Relation usage count for all subj x obj category combinations, showing label


get_relation_usage_pivot_resource Relation usage count for all subj x obj category combinations


get_relation_usage_resource All relations used plus count of associations

}
}

\keyword{datasets}
