% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchApi.r
\docType{data}
\name{SearchApi}
\alias{SearchApi}
\title{Search operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SearchApi
}
\description{
swagger.Search
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

get_autocomplete Returns list of matching concepts or entities using lexical search


get_search_entities Returns list of matching concepts or entities using lexical search

}
}

\keyword{datasets}
