% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchResult.r
\docType{data}
\name{SearchResult}
\alias{SearchResult}
\title{SearchResult Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SearchResult
}
\description{
SearchResult Class
}
\section{Fields}{

\describe{
\item{\code{start}}{}

\item{\code{facet_pivot}}{}

\item{\code{facet_counts}}{}

\item{\code{numFound}}{}
}}

\keyword{datasets}
