% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seq.r
\docType{data}
\name{Seq}
\alias{Seq}
\title{Seq Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Seq
}
\description{
Seq Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{md5checksum}}{}

\item{\code{residues}}{}

\item{\code{alphabet}}{}

\item{\code{seqlen}}{}
}}

\keyword{datasets}
