% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceFeature.r
\docType{data}
\name{SequenceFeature}
\alias{SequenceFeature}
\title{SequenceFeature Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequenceFeature
}
\description{
SequenceFeature Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{locations}}{}

\item{\code{homology_associations}}{}

\item{\code{seq}}{}
}}

\keyword{datasets}
