% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceLocation.r
\docType{data}
\name{SequenceLocation}
\alias{SequenceLocation}
\title{SequenceLocation Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequenceLocation
}
\description{
SequenceLocation Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{start}}{}

\item{\code{strand}}{}

\item{\code{phase}}{}

\item{\code{end}}{}

\item{\code{score}}{}
}}

\keyword{datasets}
