% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequencePosition.r
\docType{data}
\name{SequencePosition}
\alias{SequencePosition}
\title{SequencePosition Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequencePosition
}
\description{
SequencePosition Class
}
\section{Fields}{

\describe{
\item{\code{reference}}{}

\item{\code{position}}{}
}}

\keyword{datasets}
