% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPropertyValue.r
\docType{data}
\name{SummaryPropertyValue}
\alias{SummaryPropertyValue}
\title{SummaryPropertyValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SummaryPropertyValue
}
\description{
SummaryPropertyValue Class
}
\section{Fields}{

\describe{
\item{\code{pred}}{}

\item{\code{xrefs}}{}

\item{\code{val}}{}
}}

\keyword{datasets}
