% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynonymPropertyValue.r
\docType{data}
\name{SynonymPropertyValue}
\alias{SynonymPropertyValue}
\title{SynonymPropertyValue Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SynonymPropertyValue
}
\description{
SynonymPropertyValue Class
}
\section{Fields}{

\describe{
\item{\code{pred}}{}

\item{\code{xrefs}}{}

\item{\code{val}}{}
}}

\keyword{datasets}
