% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcript.r
\docType{data}
\name{Transcript}
\alias{Transcript}
\title{Transcript Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Transcript
}
\description{
Transcript Class
}
\section{Fields}{

\describe{
\item{\code{id}}{}

\item{\code{label}}{}

\item{\code{categories}}{}

\item{\code{types}}{}

\item{\code{description}}{}

\item{\code{consider}}{}

\item{\code{deprecated}}{}

\item{\code{replaced_by}}{}

\item{\code{synonyms}}{}

\item{\code{xrefs}}{}

\item{\code{taxon}}{}

\item{\code{exons}}{}

\item{\code{genes}}{}

\item{\code{sequence_features}}{}

\item{\code{cds}}{}
}}

\keyword{datasets}
