% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariantSet.r
\docType{data}
\name{VariantSet}
\alias{VariantSet}
\title{VariantSet Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariantSet
}
\description{
VariantSet Class
}
\section{Fields}{

\describe{
\item{\code{body}}{}

\item{\code{id}}{}

\item{\code{title}}{}

\item{\code{category_id}}{}

\item{\code{pub_date}}{}

\item{\code{category}}{}
}}

\keyword{datasets}
