% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariationsetApi.r
\docType{data}
\name{VariationsetApi}
\alias{VariationsetApi}
\title{Variationset operations}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
VariationsetApi
}
\description{
swagger.Variationset
}
\section{Fields}{

\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}}

\section{Methods}{

\describe{

delete_variant_set_item Deletes variant set


get_variant_analyze Returns list of matches


get_variant_set_item Returns a variant set


get_variant_sets_archive_collection Returns list of variant sets from a specified time period


get_variant_sets_collection Returns list of variant sets


post_variant_sets_collection Creates a new variant set


put_variant_set_item Updates a variant set

}
}

\keyword{datasets}
